Move functions to cpp file to fix linking issues
