+++ src/libasr/codegen/llvm_utils.cpp
@@ -2,6 +2,25 @@
 
 namespace LFortran {
 
+    namespace LLVM {
+
+        llvm::Value* CreateLoad(llvm::IRBuilder<> &builder, llvm::Value *x) {
+            // TODO: the type must be dereferenced:
+            return builder.CreateLoad(x->getType(), x);
+        }
+
+
+        llvm::Value* CreateGEP(llvm::IRBuilder<> &builder, llvm::Value *x, std::vector<llvm::Value *> &idx) {
+            // TODO: the type must be dereferenced:
+            return builder.CreateGEP(x->getType(), x, idx);
+        }
+
+        llvm::Value* CreateInBoundsGEP(llvm::IRBuilder<> &builder, llvm::Value *x, std::vector<llvm::Value *> &idx) {
+            // TODO: the type must be dereferenced:
+            return builder.CreateInBoundsGEP(x->getType(), x, idx);
+        }
+    }
+
     LLVMUtils::LLVMUtils(llvm::LLVMContext& context,
         llvm::IRBuilder<>* _builder):
         context(context),
